Fix parsing optional function return types
`(Int) -> Bool?` was being parsed as `((Int) -> Bool)?`
