+++ MockingbirdGenerator/Parser/Models/Parser/Tuple.swift
@@ -1,248 +1,11 @@
 //
-//  DeclaredType.swift
+//  Tuple.swift
 //  MockingbirdGenerator
 //
-//  Created by Andrew Chang on 9/1/19.
+//  Created by Andrew Chang on 3/24/20.
 //
 
 import Foundation
-
-/// Tokenizes type declarations.
-enum DeclaredType: CustomStringConvertible, CustomDebugStringConvertible, SerializableType {
-  case single(Single, optionals: String)
-  case tuple(Tuple, optionals: String)
-  
-  var isFunction: Bool {
-    switch self {
-    case let .single(single, _): return single.isFunction
-    case .tuple: return false
-    }
-  }
-  
-  var isCollection: Bool {
-    switch self {
-    case let .single(single, _): return single.isCollection
-    case .tuple: return false
-    }
-  }
-  
-  var isTuple: Bool {
-    switch self {
-    case .single: return false
-    case .tuple: return true
-    }
-  }
-  
-  var isOptional: Bool {
-    switch self {
-    case .single(_, let optionals), .tuple(_, let optionals): return !optionals.isEmpty
-    }
-  }
-  
-  var description: String {
-    switch self {
-    case let .single(single, optionals):
-      if isOptional && isFunction {
-        return "(\(single))\(optionals)"
-      } else {
-        return "\(single)\(optionals)"
-      }
-    case let .tuple(tuple, optionals): return "\(tuple)\(optionals)"
-    }
-  }
-  
-  var debugDescription: String {
-    var description: String {
-      switch self {
-      case let .single(single, optionals):
-        if isOptional && isFunction {
-          return "(\(String(reflecting: single)))\(optionals)"
-        } else {
-          return "\(String(reflecting: single))\(optionals)"
-        }
-      case let .tuple(tuple, optionals): return "\(String(reflecting: tuple))\(optionals)"
-      }
-    }
-    return "DeclaredType(\(description))"
-  }
-  
-  func serialize(with request: SerializationRequest) -> String {
-    let processOptionals: (String) -> String = { optionals in
-      guard request.options.contains(.shouldExcludeImplicitlyUnwrappedOptionals),
-        optionals.hasSuffix("!") else { return optionals }
-      return String(optionals.dropLast()) // Not possible to implicitly unwrap multiple times.
-    }
-    switch self {
-    case let .single(single, optionals):
-      if isOptional && isFunction {
-        return "(" + single.serialize(with: request) + ")" + processOptionals(optionals)
-      } else  {
-        return single.serialize(with: request) + processOptionals(optionals)
-      }
-    case let .tuple(tuple, optionals):
-      return tuple.serialize(with: request) + processOptionals(optionals)
-    }
-  }
-}
-
-extension DeclaredType {
-  init(from serialized: String, ignoreCache: Bool = false) {
-    self.init(from: serialized[...], ignoreCache: ignoreCache)
-  }
-  
-  init(from serialized: Substring, ignoreCache: Bool = false) {
-    if !ignoreCache, let primitive = Primitives.map[serialized] {
-      self = primitive
-      return
-    }
-    
-    let trimmed = serialized.trimmingCharacters(in: .whitespacesAndNewlines)[...]
-    // Handle optionals (which can be wrapped multiple times).
-    let firstOptionalIndex = trimmed.firstIndex(of: "?", excluding: .allGroups)
-      ?? trimmed.firstIndex(of: "!", excluding: .allGroups)
-      ?? trimmed.endIndex
-    var optionals = String(trimmed[firstOptionalIndex...])
-    var unwrappedType = trimmed[..<firstOptionalIndex]
-    guard let tuple = Tuple(from: unwrappedType) else {
-      while unwrappedType.hasPrefix("(")
-        && unwrappedType.hasSuffix(")")
-        && !unwrappedType.contains("->", excluding: .allGroups) {
-          unwrappedType = unwrappedType.dropFirst().dropLast()
-          
-          // Coalesce unwrapped optionals.
-          let firstOptionalIndex = unwrappedType.firstIndex(of: "?", excluding: .allGroups)
-            ?? unwrappedType.firstIndex(of: "!", excluding: .allGroups)
-            ?? unwrappedType.endIndex
-          optionals = String(unwrappedType[firstOptionalIndex...]) + optionals
-          unwrappedType = unwrappedType[..<firstOptionalIndex]
-      }
-      self = .single(Single(from: unwrappedType, ignoreCache: ignoreCache), optionals: optionals)
-      return
-    }
-    self = .tuple(tuple, optionals: optionals)
-  }
-}
-
-indirect enum Single: CustomStringConvertible, CustomDebugStringConvertible, SerializableType {
-  case generic(
-    typeName: String,
-    qualification: String,
-    genericTypes: [DeclaredType]
-  )
-  case list(element: DeclaredType)
-  case map(key: DeclaredType, value: DeclaredType)
-  case function(Function)
-  
-  var isFunction: Bool {
-    switch self {
-    case .function: return true
-    default: return false
-    }
-  }
-  
-  var isCollection: Bool {
-    switch self {
-    case .list, .map: return true
-    default: return false
-    }
-  }
-  
-  var description: String {
-    switch self {
-    case let .generic(typeName, qualification, genericTypes):
-      let qualificationPrefix = qualification + (qualification.isEmpty ? "" : ".")
-      guard !genericTypes.isEmpty else { return "\(qualificationPrefix)\(typeName)" }
-      return "\(qualificationPrefix)\(typeName)<\(genericTypes.map({ "\($0)" }).joined(separator: ", "))>"
-    case let .list(element):
-      return "[\(element)]"
-    case let .map(key, value):
-      return "[\(key): \(value)]"
-    case let .function(function):
-      return "\(function)"
-    }
-  }
-  
-  var debugDescription: String {
-    var description: String {
-      switch self {
-      case let .generic(typeName, qualification, genericTypes):
-        let qualificationPrefix = qualification + (qualification.isEmpty ? "" : ".")
-        guard !genericTypes.isEmpty else { return "\(qualificationPrefix)\(typeName)" }
-        return "\(qualificationPrefix)\(typeName)<\(genericTypes.map({ String(reflecting: $0) }).joined(separator: ", "))>"
-      case let .list(element):
-        return "[\(String(reflecting: element))]"
-      case let .map(key, value):
-        return "[\(String(reflecting: key)): \(String(reflecting: value))]"
-      case let .function(function):
-        return String(reflecting: function)
-      }
-    }
-    return "Single(\(description))"
-  }
-  
-  func serialize(with request: SerializationRequest) -> String {
-    switch self {
-    case let .generic(typeName, qualification, genericTypes):
-      let qualificationPrefix = qualification + (qualification.isEmpty ? "" : ".")
-      guard !genericTypes.isEmpty else { return "\(request.serialize(qualificationPrefix + typeName))" }
-      return "\(request.serialize(qualificationPrefix + typeName))<\(genericTypes.map({ $0.serialize(with: request) }).joined(separator: ", "))>"
-    case let .list(element):
-      return "[\(element.serialize(with: request))]"
-    case let .map(key, value):
-      return "[\(key.serialize(with: request)): \(value.serialize(with: request))]"
-    case let .function(function):
-      return "\(function.serialize(with: request))"
-    }
-  }
-}
-
-extension Single {
-  init(from serialized: Substring, ignoreCache: Bool = false) {
-    guard serialized.hasPrefix("["), serialized.hasSuffix("]") else {
-      // Handle type qualifications.
-      let components = serialized.components(separatedBy: ".", excluding: .allGroups)
-      let qualification = components[..<(components.count-1)].joined(separator: ".")
-      let baseType = components.last ?? ""
-      
-      guard let genericTypesStartIndex = baseType.firstIndex(of: "<", excluding: .allGroups),
-        let genericTypesEndIndex = baseType[baseType.index(after: genericTypesStartIndex)...]
-          .firstIndex(of: ">", excluding: .allGroups)
-        else {
-          if let function = Function(from: serialized) {
-            self = .function(function)
-          } else {
-            self = .generic(typeName: String(baseType),
-                            qualification: qualification,
-                            genericTypes: [])
-          }
-          return
-      }
-      
-      // Handle types with generic specializations.
-      let despecializedTypeName = baseType[..<genericTypesStartIndex]
-      let genericTypesRange = baseType.index(after: genericTypesStartIndex)..<genericTypesEndIndex
-      let genericTypes = baseType[genericTypesRange]
-        .components(separatedBy: ",", excluding: .allGroups)
-        .map({ DeclaredType(from: $0, ignoreCache: ignoreCache) })
-      self = .generic(typeName: String(despecializedTypeName),
-                      qualification: qualification,
-                      genericTypes: genericTypes)
-      return
-    }
-    
-    // Handle collection types.
-    let flattened = serialized.dropFirst().dropLast()
-    if let keyIndex = flattened.firstIndex(of: ":", excluding: .allGroups) { // Map type.
-      let keyTypeName = flattened[..<keyIndex].trimmingCharacters(in: .whitespacesAndNewlines)
-      let valueTypeName = flattened[flattened.index(after: keyIndex)...]
-        .trimmingCharacters(in: .whitespacesAndNewlines)
-      self = .map(key: DeclaredType(from: keyTypeName[...], ignoreCache: ignoreCache),
-                  value: DeclaredType(from: valueTypeName[...], ignoreCache: ignoreCache))
-    } else {
-      self = .list(element: DeclaredType(from: flattened, ignoreCache: ignoreCache))
-    }
-  }
-}
 
 struct Tuple: CustomStringConvertible, CustomDebugStringConvertible, SerializableType {
   let elements: [(label: String, type: DeclaredType)] // Label => Type
